- Deleted TableRowFilterSupport   - The goal is to have Analyster call the filter directly so I was able to eliminate the need for TableRowFilterSupport   - The TableFilterColumnPopup is now also called from Analyster rather than TableRowFilterSupport   - Unfortunately, due to accommodating legacy code, the filter and columnPopup are also called from LoadTables as well     - This is temporary until further refactoring can be done, and this is also a minor concern.
